appease clippy

+++ libnet/src/ioctl.rs
@@ -761,7 +761,7 @@ fn plumb_for_af(name: &str, ifflags: u64) -> Result<(), Error> {
     let arg = &mut req as *mut sys::lifreq as *mut c_char;
     str_ioctl(
         arp_fd,
-        sys::SIOCSLIFNAME as i32,
+        sys::SIOCSLIFNAME,
         arg,
         size_of::<sys::lifreq>() as c_int,
     )?;
@@ -1243,7 +1243,7 @@ fn ipmgmtd_persist(
     let mut nvl = nvpair::NvList::new_unique_names();
     nvl.insert("_ifname", ifname)?;
     nvl.insert("_aobjname", objname)?;
-    nvl.insert("_lifnum", &(lifnum as i32))?;
+    nvl.insert("_lifnum", &lifnum)?;
 
     match addr {
         Some(addr) => {
@@ -1444,7 +1444,7 @@ pub(crate) fn create_vnic(
             mac_slot: -1,
             vid: 0,
             vrid: 0,
-            af: AF_UNSPEC as i32,
+            af: AF_UNSPEC,
             flags: 0,
             ..Default::default()
         };
